Created a base eslint config
