+++ packages/eslint-config-custom/node.js
@@ -1,36 +1,4 @@
-const { resolve } = require("node:path");
-
-const project = resolve(process.cwd(), "tsconfig.json");
-
 module.exports = {
-  env: { node: true, es2020: true },
-  extends: [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended-type-checked",
-    "plugin:@typescript-eslint/stylistic-type-checked",
-    "prettier",
-    "turbo",
-    "plugin:mdx/recommended",
-  ],
-  ignorePatterns: ["dist/", "node_modules/"],
-  parser: "@typescript-eslint/parser",
-  plugins: ["@typescript-eslint"],
-  rules: {
-    semi: "error",
-    "no-empty": "warn",
-    "no-empty-function": "warn",
-    "@typescript-eslint/no-unused-vars": "warn",
-    "prefer-const": "warn",
-    "@typescript-eslint/no-empty-interface": "warn",
-  },
-  parserOptions: {
-    project,
-  },
-  settings: {
-    "import/resolver": {
-      typescript: {
-        project,
-      },
-    },
-  },
+  env: { node: true },
+  extends: ["./base"],
 };
